add to cart, changed orders functionality,
all the pages now use user id and item id to fetch data from backend. The add to cart button sends data to server regarding the user and the item to be added the items are not being actually added now. The orders now contain seller and buyer id and status is either pending or completed. Database is yet to be connnected